Improvement

+++ App/Controller/Backend.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace citymobile;
+
 use Administrator;
 use Exception;
 use Photo;
@@ -26,7 +27,7 @@ class ControllerBackend
      */
     public function home()
     {
-        if(!AdministratorManager::isConnected())
+        if (!AdministratorManager::isConnected())
             header('location: index.php?p=admin_login');
         require '../views/Backend/home.php';
     }
@@ -36,7 +37,7 @@ class ControllerBackend
      */
     public function login()
     {
-        if(AdministratorManager::isConnected())
+        if (AdministratorManager::isConnected())
             header('location: index.php?p=admin_home');
 
         if (!empty($_POST)) {
@@ -47,7 +48,7 @@ class ControllerBackend
             else {
                 $administratorManager = new AdministratorManager();
                 $errors = $administratorManager->connect($administrator);
-                if(!empty($errors))
+                if (!empty($errors))
                     require '../views/backend/login.php';
                 else
                     header('location: index.php?p=admin_home');
@@ -76,37 +77,73 @@ class ControllerBackend
                 $this->articleManager->save($article);
                 echo '<div class="container"><div class="alert alert-success">Votre article à bien été ajouté <br />> <a href="index.php?p=admin_list_articles">Afficher la liste des articles.</a></div></div>';
             }
-
         } else {
             require '../views/backend/addArticle.php';
         }
     }
 
-    public function editArticle()
+    public function updateArticle($id)
     {
-        if(!AdministratorManager::isConnected())
+        $id = (int)$id;
+        if (!AdministratorManager::isConnected())
             header('location: index.php?p=admin_login');
-        $photo = new Photo($_FILES);
-        $article = new Article($_POST);
-        if ($article->isNew())
-            throw new Exception("L'article que vous essayer de modifier n'existe pas, veuillez le créé.");
-        $article->setPhoto($photo->getName());
-        $this->articleManager->save($article);
+
+        if (!empty($_POST)) {
+
+            if(empty($_FILES['photo']['name']))
+                $photoName = $_POST['oldPhoto'];
+            else {
+                $photo = new Photo($_FILES);
+                $photo->add();
+                $photoName = $photo->getName();
+            }
+
+            $article = new Article($_POST);
+            if ($article->isNew())
+                throw new Exception("L'article que vous essayer de modifier n'existe pas, <a href='index.php?p=admin_add_article'>cliquez ici pour en crée un</a>.");
+
+            $article->setPhoto($_POST['oldPhoto']);
+            $article->setPhoto($photoName);
+            $errors = $article->errors;
+            if (!empty($errors))
+                require '../views/backend/updateArticle.php';
+            else {
+                if(!empty($_FILES['name']))
+                    $photo->add(); // Add photo on the server.
+                $this->articleManager->save($article);
+                echo '<div class="container"><div class="alert alert-success">Votre article à bien été modifié <br />> <a href="index.php?p=admin_list_articles">Afficher la liste des articles.</a></div></div>';
+            }
+        } else {
+            $article = $this->articleManager->get($id);
+            require '../views/backend/updateArticle.php';
+        }
 
         // View
     }
 
-    public function listArticles()
+    public
+    function listArticles()
     {
+        if (!AdministratorManager::isConnected())
+            header('location: index.php?p=admin_login');
         $articles = $this->articleManager->getList();
         require '../views/backend/listArticles.php';
     }
 
-    public function deleteArticle($token)
+    public
+    function deleteArticle($token)
     {
-        if(isset($token) && !empty($token)) {
-            $this->articleManager->delete($token);
+        if (isset($token) && !empty($token)) {
+            if ($this->articleManager)
+                $this->articleManager->delete($token);
+            echo '<div class="container"><div class="alert alert-success">L\'article à bien été supprimé <br />> <a href="index.php?p=admin_list_articles">Afficher la liste des articles.</a></div></div>';
         }
+    }
+
+    public function logout()
+    {
+        session_destroy();
+        header('location: index.php?p=home');
     }
